Crowd grouping and force velocity
Put in some adjustments to get the crowd to stay inside the original
defined crowd area. I'm also trying to change the crowd jumping to use a
velocity instead of a force so that we can avoid worrying about changes
to the crowd member's mass. Unfortunately, I'm not seeing the numbers I
want on the crowdie, despite the calculations I'm making for the initial
velocity seeming correct. Will try again tomorrow.